# api/openapi.yaml
components:
  schemas:
    ActiveDateRange:
      additionalProperties: false
      properties:
        from:
          type: string
        to:
          type: string
      required:
        - from
        - to
      type: object
    ActiveFilters:
      additionalProperties: false
      properties:
        date_range:
          $ref: "#/components/schemas/ActiveDateRange"
        show_hidden:
          type: boolean
        show_transfers:
          type: boolean
      required:
        - show_hidden
        - show_transfers
      type: object
    AggregateRow:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        dimension:
          type: string
        flags:
          $ref: "#/components/schemas/Flags"
        identity:
          type: string
        index:
          format: int64
          type: integer
        label:
          type: string
        period:
          $ref: "#/components/schemas/Period"
        share_tenths:
          format: int64
          type: integer
        top_category:
          type: string
        top_category_percent:
          format: int64
          type: integer
        total:
          $ref: "#/components/schemas/Money"
      required:
        - index
        - identity
        - dimension
        - label
        - count
        - total
        - share_tenths
        - flags
      type: object
    Bootstrap:
      additionalProperties: false
      properties:
        base_path:
          type: string
        canonical_url:
          type: string
        mutation_token:
          type: string
        revision:
          pattern: ^[0-9]+$
          type: string
        token_expires_at:
          format: date-time
          type: string
        version:
          type: string
      required:
        - version
        - canonical_url
        - base_path
        - revision
        - mutation_token
        - token_expires_at
      type: object
    Breadcrumb:
      additionalProperties: false
      properties:
        dimension:
          type: string
        label:
          type: string
      required:
        - dimension
        - label
      type: object
    Capability:
      additionalProperties: false
      properties:
        available:
          type: boolean
        category:
          type: string
        description:
          type: string
        id:
          type: string
        key_display:
          type: string
      required:
        - id
        - description
        - category
        - available
      type: object
    Chart:
      additionalProperties: false
      properties:
        marks:
          items:
            $ref: "#/components/schemas/ChartMark"
          type:
            - array
            - "null"
        summary:
          items:
            $ref: "#/components/schemas/Statistics"
          type:
            - array
            - "null"
      type: object
    ChartMark:
      additionalProperties: false
      properties:
        amount:
          $ref: "#/components/schemas/Money"
        identity:
          type: string
        index:
          format: int64
          type: integer
        label:
          type: string
        plot_ratio:
          format: int64
          maximum: 10000
          minimum: -10000
          type: integer
      required:
        - index
        - identity
        - label
        - amount
        - plot_ratio
      type: object
    DetailRow:
      additionalProperties: false
      properties:
        account:
          type: string
        amount:
          $ref: "#/components/schemas/Money"
        category:
          type: string
        date:
          type: string
        flags:
          $ref: "#/components/schemas/Flags"
        group:
          type: string
        identity:
          type: string
        index:
          format: int64
          type: integer
        merchant:
          type: string
      required:
        - index
        - identity
        - date
        - account
        - merchant
        - category
        - group
        - amount
        - flags
      type: object
...
    Flags:
      additionalProperties: false
      properties:
        hidden:
          type: boolean
        pending:
          type: boolean
        selected:
          type: boolean
      required:
        - selected
        - hidden
        - pending
      type: object
...
    Money:
      additionalProperties: false
      properties:
        currency:
          type: string
        decimal:
          type: string
        display:
          type: string
        minor:
          type: string
        scale:
          format: int32
          minimum: 0
          type: integer
      required:
        - minor
        - currency
        - scale
        - decimal
        - display
      type: object
    Period:
      additionalProperties: false
      properties:
        day:
          format: int64
          type: integer
        granularity:
          type: string
        month:
          format: int64
          type: integer
        year:
          format: int64
          type: integer
      required:
        - granularity
        - year
      type: object
    Problem:
      additionalProperties: false
      properties:
        code:
          type: string
        detail:
          type: string
        status:
          format: int64
          type: integer
        title:
          type: string
        type:
          default: about:blank
          format: uri
          type: string
      required:
        - title
        - status
        - detail
        - code
      type: object
    Projection:
      additionalProperties: false
      properties:
        aggregate_rows:
          items:
            $ref: "#/components/schemas/AggregateRow"
          type:
            - array
            - "null"
        api_schema_version:
          type: string
        breadcrumb_text:
          type: string
        breadcrumbs:
          items:
            $ref: "#/components/schemas/Breadcrumb"
          type:
            - array
            - "null"
        canonical_query:
          type: string
        capabilities:
          items:
            $ref: "#/components/schemas/Capability"
          type:
            - array
            - "null"
        chart:
          $ref: "#/components/schemas/Chart"
        detail_rows:
          items:
            $ref: "#/components/schemas/DetailRow"
          type:
            - array
            - "null"
        filters:
          $ref: "#/components/schemas/ActiveFilters"
        projection_schema_version:
          type: string
        selection:
          type: string
        statistics:
          items:
            $ref: "#/components/schemas/Statistics"
          type:
            - array
            - "null"
        status:
          type: string
        total_rows:
          format: int64
          type: integer
        view:
          $ref: "#/components/schemas/ViewMetadata"
        warnings:
          items:
            $ref: "#/components/schemas/Warning"
          type:
            - array
            - "null"
        window:
          $ref: "#/components/schemas/ReturnedWindow"
      required:
        - api_schema_version
        - projection_schema_version
        - canonical_query
        - selection
        - view
        - breadcrumbs
        - breadcrumb_text
        - filters
        - capabilities
        - total_rows
        - window
        - statistics
        - chart
      type: object
    ReturnedWindow:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        limit:
          format: int64
          type: integer
        offset:
          format: int64
          type: integer
      required:
        - offset
        - limit
        - count
      type: object
    Statistics:
      additionalProperties: false
      properties:
        count:
          format: int64
          type: integer
        currency:
          type: string
        in:
          $ref: "#/components/schemas/Money"
        net:
          $ref: "#/components/schemas/Money"
        out:
          $ref: "#/components/schemas/Money"
        scale:
          format: int32
          minimum: 0
          type: integer
      required:
        - currency
        - scale
        - count
        - in
        - out
        - net
      type: object
    TransitionBody:
      additionalProperties: false
      properties:
        action:
          type: string
        filters:
          $ref: "#/components/schemas/TransitionFilters"
        query:
          maxLength: 65536
          type: string
        search:
          maxLength: 2048
          type: string
        selection:
          maxLength: 1468006
          type: string
        target:
          $ref: "#/components/schemas/TransitionTarget"
        window:
          $ref: "#/components/schemas/Window"
      required:
        - query
        - action
        - window
      type: object
    TransitionDateRange:
      additionalProperties: false
      properties:
        end:
          format: date
          type: string
        start:
          format: date
          type: string
      required:
        - start
        - end
      type: object
    TransitionFilters:
      additionalProperties: false
      properties:
        date_range:
          $ref: "#/components/schemas/TransitionDateRange"
        show_hidden:
          type: boolean
        show_transfers:
          type: boolean
      required:
        - show_hidden
        - show_transfers
      type: object
    TransitionTarget:
      additionalProperties: false
      properties:
        identity:
          maxLength: 4096
          type: string
        kind:
          type: string
      required:
        - kind
        - identity
      type: object
    ViewBody:
      additionalProperties: false
      properties:
        query:
          maxLength: 65536
          type: string
        selection:
          maxLength: 1468006
          type: string
        window:
          $ref: "#/components/schemas/Window"
      required:
        - query
        - window
      type: object
    ViewMetadata:
      additionalProperties: false
      properties:
        grouping:
          type: string
        mode:
          type: string
        search:
          type: string
        sort_direction:
          type: string
        sort_field:
          type: string
        time_granularity:
          type: string
      required:
        - mode
        - grouping
        - time_granularity
        - sort_field
        - sort_direction
      type: object
    Warning:
      additionalProperties: false
      properties:
        code:
          type: string
        detail:
          type: string
      required:
        - code
        - detail
      type: object
    Window:
      additionalProperties: false
      properties:
        limit:
          default: 200
          format: int64
          maximum: 400
          minimum: 1
          type: integer
        offset:
          default: 0
          format: int64
          maximum: 1000000
          minimum: 0
          type: integer
      required:
        - offset
        - limit
      type: object
info:
  title: moneyflow
  version: "1"
openapi: 3.1.0
paths:
  /api/v1/bootstrap:
    get:
      operationId: bootstrap
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
          description: OK
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
      summary: Issue browser-memory mutation configuration
...
  /api/v1/view:
    post:
      operationId: projectView
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projection"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
      summary: Project one stateless read-only view
  /api/v1/view/transition:
    post:
      operationId: transitionView
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionBody"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Projection"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Bad Request
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Conflict
        "413":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Request Entity Too Large
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Unprocessable Entity
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal Server Error
      summary: Apply one stateless read-only view transition
servers:
  - url: /